refactor: Switch tests to ava.
