+++ src/tests/stdlib.spec.ts
@@ -1,0 +1,22 @@
+import test from 'ava';
+import { runCode } from './util';
+
+test('test stdlib console.log', async t => {
+  await runCode(
+    `
+    let globalstr = "globalstr\n";
+
+    function main(): number {
+        let localstr = "localstr\n";
+        console.log(globalstr);
+        console.log(localstr);
+        console.log("%s", globalstr);
+        console.log("%s", localstr);
+        console.log("%d\n", 42);
+        return 0;
+    }
+    `
+  );
+
+  t.pass();
+});
