--- tests/ts/enum/initializer.ts
@@ -1,9 +1,0 @@
-enum Test {
-  a = 100,
-  b = 101,
-  c = 102
-}
-
-function main(): number {
-  return Test.a;
-}
