--- tests/ts/enum/iota.ts
@@ -1,9 +1,0 @@
-enum Test {
-  a = 100,
-  b,
-  c = 103
-}
-
-function main(): number {
-  return Test.b;
-}
