--- tests/ts/enum/return.ts
@@ -1,8 +1,0 @@
-enum Test {
-  a,
-  b
-}
-
-function main(): Test {
-  return Test.a;
-}
