--- tests/ts/enum/var.ts
@@ -1,12 +1,0 @@
-enum Test {
-  a,
-  b
-}
-
-function test(t: Test): number {
-  return t;
-}
-
-function main(): number {
-  return test(Test.b);
-}
