--- tests/ts/exit_0.ts
@@ -1,3 +1,0 @@
-function main(): number {
-    return 0;
-}
