--- tests/ts/exit_1.ts
@@ -1,3 +1,0 @@
-function main(): number {
-    return 1;
-}
