--- tests/ts/global_var.ts
@@ -1,7 +1,0 @@
-let a: number = 1;
-let b: number[] = [2, 3];
-
-function main(): number {
-    let c = 4;
-    return a + b[0] + c;
-}
