--- tests/ts/object/return_number.ts
@@ -1,13 +1,0 @@
-const Test = {
-  a: 1,
-  d: 12
-};
-
-function main(): number {
-  const Test = {
-    a: 1,
-    c: 10
-  };
-
-  return Test.c;
-}
