--- tests/ts/object/return_string.ts
@@ -1,16 +1,0 @@
-const Test = {
-  a: 1
-};
-
-function main(): number {
-  const Test = {
-    a: 1,
-    c: 'str'
-  };
-
-  if (Test.c === 'str') {
-    return 0;
-  }
-
-  return 1;
-}
