--- tests/ts/operator/arithmetic_operations.ts
@@ -1,10 +1,0 @@
-function main(): number {
-    let n: number = 0;
-    n = n + 10; // 10
-    n = n - 2;  // 8
-    n = n * 2;  // 16
-    n = n / 4;  // 4
-    n = n * 10; // 40
-    n = n % 17; // 6
-    return n;
-}
