--- tests/ts/operator/compound_assignment_const.ts
@@ -1,16 +1,0 @@
-// *=, /=, %=, +=, -=, <<=, >>=, >>>=, &=, ^=, |=
-function main(): number {
-  let a: number = 1024;
-  a *= 2;
-  a /= 2;
-  a %= 2;
-  a += 2;
-  a -= 2;
-  a <<= 2;
-  a >>= 2;
-  a >>>= 2;
-  a &= 2;
-  a *= 2;
-  a |= 2;
-  return a;
-}
