--- tests/ts/operator/compound_assignment_var.ts
@@ -1,17 +1,0 @@
-// *=, /=, %=, +=, -=, <<=, >>=, >>>=, &=, ^=, |=
-function main(): number {
-  let a: number = 1024;
-  let b: number = 2;
-  a *= b;
-  a /= b;
-  a %= b;
-  a += b;
-  a -= b;
-  a <<= b;
-  a >>= b;
-  a >>>= b;
-  a &= b;
-  a *= b;
-  a |= b;
-  return a;
-}
