--- tests/ts/operator/number.ts
@@ -1,3 +1,0 @@
-function main(): number {
-    return 0x0a + 10; // 20
-}
