--- tests/ts/operator/postfix_unary.ts
@@ -1,9 +1,0 @@
-function main(): number {
-    let a: number = 0;
-    a++;
-    a++;
-    a++;
-    a--;
-    let b: number = a--;
-    return b;
-}
