--- tests/ts/statement/do.ts
@@ -1,9 +1,0 @@
-function main(): number {
-    let s = 0;
-    let i = 0;
-    do {
-        s += i;
-        i += 1;
-    } while (i <= 10)
-    return s; // 55
-}
