--- tests/ts/statement/while.ts
@@ -1,7 +1,0 @@
-function main(): number {
-    let s = 0;
-    while (s != 10) {
-        s++;
-    }
-    return s; // 20
-}
