--- tests/ts/type_inference.ts
@@ -1,5 +1,0 @@
-function main(): number {
-    let a = 10; // Skip the annotation type
-    let b: number = 20;
-    return a + b;
-}
